Replace donation detail page navigation with modal overlay on donor profile
Clicking a donation row from the donor profile now opens DonationViewModal
as an overlay (matching the donations page behavior) instead of navigating
to a separate /donations/:id route. Removes the dedicated details page,
useDonationDetail hook, and showOverlay/closeOnSave props that were added
to support the old approach.
